+++ src/vue/modules/create-asset-form/components/information-step-form.vue
@@ -29,10 +29,7 @@
           @blur="touchField('form.code')"
           name="create-asset-code"
           :label="'create-asset-form.code-lbl' | globalize"
-          :error-message="getFieldErrorMessage(
-            'form.code',
-            { length: CODE_MAX_LENGTH }
-          )"
+          :error-message="getFieldErrorMessage('form.code')"
           :maxlength="CODE_MAX_LENGTH"
         />
       </div>
@@ -139,7 +136,7 @@ import { DocumentContainer } from '@/js/helpers/DocumentContainer'
 
 import { CreateAssetRequest } from '../wrappers/create-asset-request'
 
-import { required, amountRange, maxLength } from '@validators'
+import { required, amountRange, maxLength, assetCode } from '@validators'
 
 import config from '@/config'
 
@@ -147,7 +144,6 @@ const EVENTS = {
   submit: 'submit',
 }
 
-const CODE_MAX_LENGTH = 16
 const NAME_MAX_LENGTH = 255
 
 export default {
@@ -172,7 +168,6 @@ export default {
     MAX_AMOUNT: config.MAX_AMOUNT,
     ASSET_POLICIES,
     DOCUMENT_TYPES,
-    CODE_MAX_LENGTH,
     NAME_MAX_LENGTH,
   }),
 
@@ -185,7 +180,7 @@ export default {
         },
         code: {
           required,
-          maxLength: maxLength(CODE_MAX_LENGTH),
+          assetCode,
         },
         maxIssuanceAmount: {
           required,